backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# AI/ML Models
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
sacremoses==0.1.1
joblib==1.3.2
scikit-learn==1.3.2

# Web Scraping
newspaper3k==0.2.8
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
opencv-python==4.8.1.78

# Language Processing
langdetect==1.0.9
indic-transliteration==2.3.64
groq==0.9.0

# Email
python-dotenv==1.0.0
aiosmtplib==3.0.1
email-validator==2.1.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
